Basic update to 2021 info
Because I don't know React and don't have capacity to learn it right now, I commented over the old code for speakers, etc so it could be used as a template later when we get the updated information. Otherwise, updated dates and text to say 2021 and discuss this year's conference.